cve-check: annotate CVEs during analysis
Add status information for each CVE under analysis.

Previously the information passed between different function of the
cve-check class included only tables of patched, unpatched, ignored
vulnerabilities and the general status of the recipe.

The VEX work requires more information, and we need to pass them
between different functions, so that it can be enriched as the
analysis progresses. Instead of multiple tables, use a single one
with annotations for each CVE encountered. For example, a patched
CVE will have:

{"abbrev-status": "Patched", "status": "version-not-in-range"}

abbrev-status contains the general status (Patched, Unpatched,
Ignored and Unknown that will be added in the VEX code)
status contains more detailed information that can come from
CVE_STATUS and the analysis.

Additional fields of the annotation include for example the name
of the patch file fixing a given CVE.

The side-effect of this change is that all entries from CVE_STATUS
are available in the result file. That includes entries from
the optional file cve-extra-exclusions.inc even if they might have
no link with the recipe (apply to a different package). This will
be fixed by moving all entries from that file to appropriate recipes.

>From now on, CVE_STATUS should be added directly in the recipe file
or in include files added only to affected recipes.

(From OE-Core rev: 8cbfbcb6146cb79ec254fe415a0e93adbe269738)

Signed-off-by: Marta Rybczynska <marta.rybczynska@syslinbit.com>
Signed-off-by: Samantha Jalabert <samantha.jalabert@syslinbit.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>